refactors from show and tell
