+++ src/Domain/Groups/Group.cs
@@ -3,29 +3,29 @@ namespace Linn.Authorisation.Domain.Groups
     using System.Collections.Generic;
     using System.Linq;
 
-    public class Group
+    public class Group : Entity
     {
         public Group(string name, bool active)
         {
             this.Name = name;
             this.Active = active;
-            this.Members = new List<GroupMember>();
+            this.Members = new List<Member>();
         }
 
         public string Name { get; set; }
 
         public bool Active { get; set; }
 
-        public IList<GroupMember> Members { get; private set; }
+        public IList<Member> Members { get; private set; }
 
         public void AddIndividualMember(string uri, string addedBy)
         {
-            this.Members.Add(new IndividualGroupMember(uri, addedBy));
+            this.Members.Add(new IndividualMember(uri, addedBy));
         }
 
         public void AddGroupMember(Group group, string addedBy)
         {
-            this.Members.Add(new GroupGroupMember(group, addedBy));
+            this.Members.Add(new GroupMember(group, addedBy));
         }
 
         public IEnumerable<string> MemberUris()
