+++ src/Domain/Groups/Member.cs
@@ -1,0 +1,14 @@
+namespace Linn.Authorisation.Domain.Groups
+{
+    using System;
+    using System.Collections.Generic;
+
+    public abstract class Member : Entity
+    {
+        public DateTime DateAdded { get; set; }
+
+        public string AddedByUri { get; set; }
+
+        public abstract IEnumerable<string> MemberUris();
+    }
+}
